requirements.txt
attr==0.3.2
beautifulsoup4==4.12.3
chromadb==0.5.2
ConfigParser==7.1.0
cryptography==44.0.0
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
gunicorn==20.1.0
PyMuPDF==1.21.1
Flask==3.1.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.5
keyring==25.5.0
langchain==0.3.13
langchain_community==0.3.13
langchain_core==0.3.28
langchain_openai==0.2.14
lockfile==0.12.2
numpy==1.26.4
openai==1.58.1
openpyxl==3.0.10
pandas==1.5.3
pdfplumber==0.11.4
Pillow==11.0.0
protobuf==5.29.2
psycopg2==2.9.9
pyOpenSSL==24.3.0
pypdf==5.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
railroad==0.5.0
selenium==4.27.1
Sphinx==8.1.3
SQLAlchemy==2.0.36
thread==2.0.5
tornado==6.4.2
trove_classifiers==2024.10.21.16
truststore==0.10.0
urllib3_secure_extra==0.1.0
Werkzeug==3.1.3
xlrd==2.0.1

